requirements.txt
boto3==1.17.106
botocore==1.20.106
box2d-py==2.3.8
colour==0.1.5
dm-env==1.5
dm-tree==0.1.6
glfw==2.1.0
gtimer==1.0.0b5
gym==0.18.3
imageio==2.9.0
imageio-ffmpeg==0.4.5
keras==2.6.0
Keras-Preprocessing==1.1.2
lightly==1.1.15
lightly-utils==0.0.1
lxml==4.6.3
matplotlib==3.4.2
more-itertools==4.2.0
pandas==1.3.4
Pillow==8.2.0
pybullet==3.1.7
pygame==2.0.1
pyOpenSSL==19.0.0
scikit-learn==0.24.2
scipy==1.7.0
seaborn==0.11.2
sk-video==1.1.10
sklearn==0.0
tensorboard==2.6.0
tensorboardX==2.4
tensorflow==2.6.0
torch==1.9.0
torchvision==0.10.0
tqdm==4.61.2
wandb==0.11.2
yarl==1.6.3
